requirements.txt
aioredis==2.0.0
alipay-sdk-python==3.3.398
aliyun-python-sdk-core-v3==2.13.11
amqp==2.6.1
appnope==0.1.2
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
billiard==3.6.3.0
bitmex==0.2.2
bravado==11.0.3
bravado-core==5.17.0
captcha==0.3
celery==4.3.1
celery-redbeat==2.0.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.4
coverage==5.2.1
crypto==1.4.1
cryptography==3.2.1
decorator==4.4.2
Django==2.2.7
django-basicauth==0.5.3
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==3.6.0
django-db-geventpool==3.1.0
django-grpc==1.0.6
django-guardian==2.0.0
django-mdeditor==0.1.18
django-ranged-response==0.2.0
django-ratelimit==2.0.0
django-redis==4.10.0
django-simple-captcha==0.5.14
django-simpleui==2021.1.1
dnspython==2.0.0
empyrical==0.5.3
eventlet==0.25.1
flower==0.9.5
geoip2==2.9.0
gevent==1.4.0
gnupg==2.3.1
greenlet==0.4.17
grpcio==1.39.0
grpcio-tools==1.39.0
gunicorn==19.9.0
humanize==3.11.0
idna==2.10
importlib-metadata==3.1.1
iniconfig==1.1.1
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
JSON-log-formatter==0.3.0
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
kombu==4.6.4
lxml==4.6.2
Markdown==3.3.4
maxminddb==2.0.3
mock==3.0.5
monotonic==1.5
more-itertools==8.8.0
msgpack==1.0.2
mysqlclient==2.0.1
Naked==0.1.31
ndg-httpsclient==0.5.1
ntplib==0.3.4
numpy==1.19.4
packaging==21.0
pandas==1.1.4
pandas-datareader==0.9.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.13.0
psutil==5.8.0
psycogreen==1.0.1
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pusher==2.1.4
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.9
Pygments==2.7.2
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyotp==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
python-dateutil==2.8.0
python-gnupg==0.4.7
pytz==2020.4
PyYAML==5.4.1
qiniu==7.2.6
qrcode==6.1
raven==6.4.0
redis==3.3.8
requests==2.25.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
rsa==4.7.2
scipy==1.5.4
shellescape==3.8.1
simplejson==3.17.3
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tenacity==8.0.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.2
uvloop==0.14.0
vine==1.3.0
wcwidth==0.2.5
webcolors==1.11.1
zipp==3.4.0
